Re-implement bip39
