--- .yarn/cache/@types-node-npm-11.11.6-40abad0842-075f1c011c.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:954ce971817fd8cedea69bf3799493b74563b2c846ad37d41c9468e877bd625a
-size 99178
