--- .yarn/cache/bip39-npm-3.0.4-7c69c9182f-79ce1600a0.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4effdee56187e2170492c8aef4184e8bcab1b81cd477a358c75903550e6d5172
-size 82326
